ci: rename ci/ublk_zoned.sh as ci/add_ublk_dev.sh
Now this script is for creating all supported ublk devices, so
rename it as add_ublk_dev.sh

Meantime start to run tests in block group, then increase timeout
to 60min.

Also install rublk with --features=zoned, so that we can cover zoned
test.

Exclude block/012 because it may fail for ublk-null, which is expected
because ublk-null doesn't save any data.

Exclude block/007 because it takes too long(20min) in github runner.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>