rublk: loop: support to pass relative path for loop
daemonize::Daemonize() will run the target code in new working
directory, so parent directory info is lost.

Pass parent directory to ublk_add_loop() for building the absolute
path.

The way is safe because we always run in same user session, and
fs open() will check if the file can be opened successfully.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>