zoned: pre-allocate space when writing from zone start
We use io_uring for zoned IO emulation, and io_uring may fallback to
io-wq always if FS needs to allocate space for each IO.

So pre-allocate zone space when writing to seq-zone beginning.

This way can avoid io-wq fallback in the following write on this zone.

sequential zone write(fio --zonemode=zbd) perf is improved a lot with this change.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>