loop: add discard/write_zeroes support
This patch adds support for discard and write_zeroes operations to the loop
target.

The `fallocate` syscall with `FALLOC_FL_PUNCH_HOLE` and
`FALLOC_FL_ZERO_RANGE` is used to implement these operations efficiently.

A new `--no-discard` command-line option is introduced to disable this
feature if needed. The device parameters are updated to reflect whether
discard is supported.

Add one test case for verifying discard.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>